separate posit docker from the inputs and put all of that in a docking.openeye module
